many addings, updated visual for main screen

+++ WeatherController/BackGround/ScenarioStructures/SettingStructure.swift
@@ -9,13 +9,13 @@ import Foundation
 
 struct SettingStructure : Codable {
     
-    var at_home: Int64
-    var co2: Int64
+    var at_home: Int
+    var co2: Int
     var dont_use: [Int]?
-    var hum: Int64
+    var hum: Int
     var must_use: [Int]?
-    var mute: Int64
-    var temp: Int64
+    var mute: Int
+    var temp: Int
     var time: String?
     
 }
